Refactor CustomRole controller and fix function management issues
Split customrole_controller.go into per-domain files (roles, grants,
functions) with reconcileOnHost as a thin orchestrator. Each domain now
owns its reconcile and cleanup logic. resolveTargetDatabases centralises
database list resolution and pre-fetches allUserDatabases once.

Fix postgres database handling: grants are never applied to postgres,
only functions. Orphaned postgres functions are now cleaned up when
spec.databases is removed (widening scope).

Fix owningRole changes being silently ignored: SyncDatabaseFunctions
now detects owner mismatches, drops the old function as its current
owner, recreates under the new owner, and follows up with ALTER
FUNCTION OWNER TO for idempotent enforcement.

Replace SET ROLE/RESET ROLE multi-statement execs with execWithRole,
which uses BEGIN/SET LOCAL ROLE/COMMIT so the role always resets on
error. Use a random dollar-quote tag to prevent body collision. Lift
desiredFunctionKey to package scope. Export UntargetableDatabases as
the shared system-database filter.
